Smoke/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
beautifulsoup4==4.12.2
browserstack-local==1.2.7
browserstack-sdk==1.18.6
bs4==0.0.1
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
chromedriver==2.24.1
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.8
docopt==0.6.2
dukpy==0.3.0
dulwich==0.21.7
execnet==2.0.2
fastjsonschema==2.19.1
filelock==3.13.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==7.0.1
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.2.0
msgpack==1.0.7
outcome==1.3.0
packaging==23.2
percy-appium-app==2.0.2
pexpect==4.9.0
Pillow==10.1.0
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==3.11.0
playwright==1.38.0
pluggy==1.3.0
psutil==5.9.6
ptyprocess==0.7.0
pycparser==2.21
pyee==9.0.4
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-SystemConfiguration==10.0
PyPAC==0.16.4
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-base-url==2.0.0
pytest-playwright==0.4.3
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
python-slugify==8.0.1
PyYAML==6.0.1
rapidfuzz==3.6.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
selenium==4.14.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
text-unidecode==1.3
tldextract==5.1.1
tomlkit==0.12.3
trio==0.22.2
trio-websocket==0.11.1
trove-classifiers==2024.1.8
typing_extensions==4.8.0
urllib3==2.0.6
virtualenv==20.25.0
wsproto==1.2.0
xattr==0.10.1
yarg==0.1.9
zipp==3.17.0
